Dependencies as Submodules
This should make it easier to build...
atm we still need to build each submodule on its own but after that it should work.
Note: assimp has an config.h that need to be relocated to the other header files.
Note: build every Submodule into a sub Dir called "build"